Port performance improvements to main diff view
